# perf: optimize tree processing