--- a/pkg/phlaredb/schemas/v1/profiles.go
+++ b/pkg/phlaredb/schemas/v1/profiles.go
@@ -325,12 +325,12 @@ type Samples struct {
 
 func NewSamples(size int) Samples {
 	return Samples{
-		StacktraceIDs: make([]uint32, 0, size),
-		Values:        make([]uint64, 0, size),
+		StacktraceIDs: make([]uint32, size),
+		Values:        make([]uint64, size),
 	}
 }
 
-func NewSamplesFromMap(m map[uint32]int64) Samples {
+func NewSamplesFromMap(m map[uint32]uint64) Samples {
 	s := Samples{
 		StacktraceIDs: make([]uint32, len(m)),
 		Values:        make([]uint64, len(m)),
@@ -339,7 +339,7 @@ func NewSamplesFromMap(m map[uint32]int64) Samples {
 	for k, v := range m {
 		if k != 0 && v > 0 {
 			s.StacktraceIDs[i] = k
-			s.Values[i] = uint64(v)
+			s.Values[i] = v
 			i++
 		}
 	}
@@ -452,41 +452,6 @@ func (s Samples) Sum() uint64 {
 		sum += v
 	}
 	return sum
-}
-
-// TODO(kolesnikovae): Consider map alternatives.
-
-// SampleMap is a map of partitioned samples structured
-// as follows: partition => stacktrace_id => value
-type SampleMap map[uint64]map[uint32]int64
-
-func (m SampleMap) Partition(p uint64) map[uint32]int64 {
-	s, ok := m[p]
-	if !ok {
-		s = make(map[uint32]int64, 128)
-		m[p] = s
-	}
-	return s
-}
-
-func (m SampleMap) AddSamples(partition uint64, samples Samples) {
-	p := m.Partition(partition)
-	for i, sid := range samples.StacktraceIDs {
-		p[sid] += int64(samples.Values[i])
-	}
-}
-
-func (m SampleMap) WriteSamples(partition uint64, dst *Samples) {
-	p, ok := m[partition]
-	if !ok {
-		return
-	}
-	dst.StacktraceIDs = dst.StacktraceIDs[:0]
-	dst.Values = dst.Values[:0]
-	for k, v := range p {
-		dst.StacktraceIDs = append(dst.StacktraceIDs, k)
-		dst.Values = append(dst.Values, uint64(v))
-	}
 }
 
 const profileSize = uint64(unsafe.Sizeof(InMemoryProfile{}))